pyproject.toml
[project]
name = "llm-fmt"
version = "0.1.0"
description = "Token-efficient data format converter for LLM contexts"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.14"
authors = [{ name = "llm-fmt contributors" }]
keywords = ["llm", "tokens", "json", "toon", "yaml", "ai", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "click>=8.1",
    "orjson>=3.10",
    "pyyaml>=6.0",
    "dpath>=2.2",
]

[project.optional-dependencies]
analysis = ["tiktoken>=0.7"]

[project.scripts]
llm-fmt = "llm_fmt.cli:main"

[project.urls]
Homepage = "https://github.com/youruser/llm-fmt"
Repository = "https://github.com/youruser/llm-fmt"
Issues = "https://github.com/youruser/llm-fmt/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_fmt"]

[dependency-groups]
dev = [
    "pre-commit~=4.0",
    "yamlfix~=1.17",
    "ruff~=0.11",
    "hatchling~=1.18",
    "hatch~=1.14",
    "basedpyright~=1.29",
]
test = [
    "pytest~=8.3",
    "pytest-mock~=3.14",
    "pytest-cov~=5.0",
    "pytest-bdd~=7.1",
    "pytest-timeout~=2.4",
]

[tool.uv]
default-groups = ["dev", "test"]

[tool.mypy]
python_version = "3.14"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=llm_fmt"
